Fix inconsistent Version and ID data-types

+++ model/event.go
@@ -13,7 +13,7 @@ type Event struct {
 	Action string `json:"action"`
 
 	// AggregateID is the ID of aggregate responsible for consuming event.
-	AggregateID int `json:"aggregate_id"`
+	AggregateID int8 `json:"aggregate_id"`
 
 	// Data is the data contained by event,
 	Data string `json:"data"`
@@ -30,9 +30,9 @@ type Event struct {
 	// Version is the version for events as processed for aggregate-projection.
 	// This is incremented by the aggregate itself each time an aggregate
 	// projection is updated.
-	Version int `json:"version"`
+	Version int64 `json:"version"`
 
 	// Year bucket is the year in which the event was generated.
 	// This is used as the partitioning key.
-	YearBucket int `json:"year_bucket"`
+	YearBucket int16 `json:"year_bucket"`
 }
